requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-community
langchain-google-genai==2.0.0
langchain-chroma
chromadb
lark